trysuperwise/pyproject.toml
[tool.poetry]
name = "trysuperwise"
version = "0.1.4"
description = "Library to start creating superwise enabled apps"
authors = ["Yash Agarwal <yagarwal1307@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
